13 columns
renamed to heat_map.py
using reprojection at begining of heat_map.py
choosing output folder does not work
